requirements.txt
allure-pytest==2.8.29
allure-python-commons==2.8.29
apipkg==1.5
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
et-xmlfile==1.1.0
execnet==1.9.0
idna==3.2
iniconfig==1.1.1
jmespath==0.10.0
jsonpath==0.82
openpyxl==3.0.7
packaging==21.0
pluggy==0.13.1
py==1.10.0
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==6.2.4
pytest-assume==2.4.3
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-repeat==0.9.1
pytest-rerunfailures==10.1
pytest-xdist==2.3.0
PyYAML==5.3.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
webdriver-manager==3.2.2
